+++ include/avif/avif.h
@@ -403,6 +403,16 @@ typedef struct avifDecoder
     uint32_t containerWidth;
     uint32_t containerHeight;
 
+    // The bit depth as reported by the AVIF container, if any. There is no guarantee
+    // this matches the decoded images; it is merely reporting what is independently offered
+    // from the container's boxes.
+    // * If decoding an "item" and the item is associated with an av1C property,
+    //   it will use the box's depth flags.
+    // * Else if decoding tracks and there is a SampleDescriptionBox of type av01 containing an av1C box,
+    //   it will use the box's depth flags.
+    // * Else it will be set to 0.
+    uint32_t containerDepth;
+
     // stats from the most recent read, possibly 0s if reading an image sequence
     avifIOStats ioStats;